CallAutomationSystem/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-auth>=2.40.2",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "gunicorn>=23.0.0",
    "pandas>=2.2.3",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.41",
    "twilio>=9.6.1",
    "werkzeug>=3.1.3",
]
